fix: move zerolog to main file
